+++ Omoni/Domain/Protocols/Repositories/ItemRepository.swift
@@ -3,6 +3,8 @@ import Foundation
 @MainActor
 protocol ItemRepository {
     func fetchItems(forItemListId itemListId: UUID) async throws -> [SDItem]
+    func fetchItems(forGroupId groupId: UUID) async throws -> [SDItem]
+    func fetchItems(forCategoryId categoryId: UUID) async throws -> [SDItem]
     func createItem(
         description: String,
         amount: Decimal,